+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `translate`, `scale` and `rotate`, thanks by the new [`UiTransform`] component.
   - It's possible to use `transform` as a shorthand property of the above ones.
 - Support for custom time support. See https://github.com/eckz/bevy_flair/blob/main/examples/animations.rs.
+- New `InlineCssStyleSheetParser` that allows parsing stylesheets from strings directly. Fixes [#8](https://github.com/eckz/bevy_flair/issues/8) 
 
 [`<gradient>`]: https://developer.mozilla.org/en-US/docs/Web/CSS/gradient